% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRedcapReport.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Download Data Report from RedCap}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{The name of the environment variable containing the RedCap API token.}

\item{redcapUrl}{The URL to the RedCap API endpoint.}

\item{redcapReportId}{The ID of the RedCap report to retrieve.}
}
\value{
A tibble containing the data from the RedCap report.
}
\description{
This function retrieves a data report from a RedCap project using API credentials stored
in the user's environment file. It requires the name of the environment variable holding
the API token, the RedCap API URL, and the report ID.
}
\examples{
# Set up in .REnviron: REDCAP_API_TOKEN="6189879441F5C29A25245880677488BF"
downloadRedcapReport("REDCAP_API_TOKEN", "https://redcap.emory.edu/api/", "46524")

}
