% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBernoulli.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Calculate the Maximum Likelihood Estimator for Parameter p of Bernoulli Distribution}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{A numeric vector of binary data (0s and 1s) where 1 represents success and 0 represents failure.}
}
\value{
The value of p that maximizes the log-likelihood of the data.
}
\description{
This function performs a grid-based search to find the value of p (probability of success)
that maximizes the log-likelihood of observing the given binary data assuming a Bernoulli
distribution. The search considers p values from 0 to 1 in increments of 0.001.
}
\examples{
test_data = c(1,0,0,0,1,1,1)
logLikBernoulli(test_data)

}
