% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcApprox.R
\name{pcApprox}
\alias{pcApprox}
\title{Approximate Data Using Principal Components}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{A numeric matrix or data frame.}

\item{npc}{The number of principal components to use for the approximation.}
}
\value{
A numeric matrix of the approximated data, rescaled and recentered.
}
\description{
This function performs PCA on the given data and uses the specified number of principal components
to reconstruct and approximate the data. The reconstructed data is rescaled and centered to match
the original data's scale and center.
}
\examples{
data_matrix = matrix(rnorm(100), ncol=10)
approx_data = pcApprox(data_matrix, 5)

}
