% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize Variable Names in a DataFrame}
\usage{
standardizeNames(data, case_type = "small_camel")
}
\arguments{
\item{data}{A dataframe or tibble whose variable names need to be standardized.}

\item{case_type}{The type of case conversion to apply, default is "small_camel".}
}
\value{
A dataframe or tibble with standardized variable names.
}
\description{
This function uses \code{janitor::make_clean_names} to clean and standardize variable names
of a dataframe to the specified case format using \code{dplyr::rename_with}. By default,
the names are converted to 'small_camel' case.
}
\examples{
data = tibble::tibble(`First name` = c("Alice", "Bob"), `Last name` = c("Smith", "Jones"))
standardized_data = standardizeNames(data)

}
