% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Reverse Scaling of a Numeric Vector}
\usage{
unscale(x)
}
\arguments{
\item{x}{A scaled numeric vector.}
}
\value{
A numeric vector with the scaling reversed, returning it to its original scale.
}
\description{
This function reverses the scaling (standardization) applied to a numeric vector.
It uses the attributes 'scaled:center' and 'scaled:scale' saved during the scaling
to revert the vector to its original form.
}
\examples{
original = c(1, 2, 3, 4, 5)
scaled = scale(original)
unscale(scaled) # should return c(1, 2, 3, 4, 5)

}
